Convert script to installable package (closes #11)
Restructure the project into a proper Python package suitable for PyPI:

- Add domain_security_analyzer/ package: __init__, __version__, __main__,
  analyzer (core logic moved from domain_analyzer.py), and cli (argparse
  interface with --help/--version plus the friendly dependency check).
- Replace domain_analyzer.py with a thin backward-compatible shim so
  `python domain_analyzer.py ...` and `from domain_analyzer import
  DomainAnalyzer` keep working.
- Add pyproject.toml (setuptools build, metadata/classifiers/keywords/URLs,
  dynamic version, `domain-analyzer` console entry point) and MANIFEST.in.
- Ignore build/dist/egg-info artifacts.

Verified: pip install -e ., `from domain_security_analyzer import
DomainAnalyzer`, `domain-analyzer --help/--version`, `python -m
domain_security_analyzer`, end-to-end 29-column CSV output, and
`python -m build` + `twine check` both pass.
