+++ .gitignore
@@ -1,2 +1,8 @@
 __pycache__/
 *.pyc
+
+# Packaging / build artifacts
+build/
+dist/
+*.egg-info/
+*.egg
