+++ MANIFEST.in
@@ -1,0 +1,7 @@
+include README.md
+include LICENSE
+include CHANGELOG.md
+include requirements.txt
+recursive-include docs *.md
+recursive-include examples *
+include scripts/*.py
