+++ domain_security_analyzer/__main__.py
@@ -1,0 +1,6 @@
+"""Enable ``python -m domain_security_analyzer``."""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
